Update OCC '19 S2 - Rimuru's Number Game.java

+++ Uncategorized/OCC '19 S2 - Rimuru's Number Game.java
@@ -6,10 +6,10 @@ public class RimurusNumberGame {
 		
 		long n = Long.parseLong(br.readLine());
 		
-		long ans = 0;
+		int ans = 0;
 		// The upper bound '524288' comes from the fact that 10,000,000,000,000,000,000 in binary is 524288 in decimal
-		for (long i = 2; i <= 524288; i ++) {
-			String binary = Long.toBinaryString(i);
+		for (int i = 2; i <= 524288; i ++) {
+			String binary = Integer.toBinaryString(i);
 			long temp = 0;
 			for (int j = 1; j < binary.length(); j ++) {
 				if (binary.charAt(j) == '1') temp += 3 * Math.pow(10, binary.length() - j - 1);
